app/requirements.txt
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
fastapi==0.101.1
FastAPI-SQLAlchemy==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
pydantic==2.2.1
pydantic_core==2.6.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
